Refactor IOMapOTBM: Extract Item OTBM serialization logic and Door logic
- Extracted `Item` (and subclasses) OTBM serialization and deserialization methods from `source/io/iomap_otbm.cpp` to a new file `source/io/item_otbm_io.cpp`.
- Moved `Door` business logic methods (`getDoorType`, `isRealDoor`, `getDoorID`, `setDoorID`) from `source/io/iomap_otbm.cpp` to `source/game/complexitem.cpp`.
- Removed unused `reform` function from `source/io/iomap_otbm.cpp`.
- Updated `source/CMakeLists.txt` to include `source/io/item_otbm_io.cpp`.
- This change decouples `Item` serialization details from the main Map I/O logic and ensures business logic resides in the appropriate model class.
